csv: generation type conversion(usize to u64)
On 32-bit targets (like WebAssembly, which RustPython supports), usize is 32-bit.
Csv with over 32-bit rows can raise panic on these platform.
type conversion from usize to u64.

+++ crates/stdlib/src/csv.rs
@@ -962,7 +962,7 @@ mod _csv {
         skipinitialspace: bool,
         delimiter: u8,
         line_num: u64,
-        generation: usize,
+        generation: u64,
     }
 
     #[pyclass(no_attr, module = "_csv", name = "reader", traverse)]
